requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
selenium==3.141.0
urllib3==1.25.9
